+++ tests/guile/ares/evaluation/test-utils.scm
@@ -15,6 +15,13 @@
      (const default-value))
     operation)))
 
+(define (valid-source? source)
+  (and (list? source)
+       (assq 'line source)
+       (assq 'column source)
+       (let ((file (assq 'file source)))
+         (or (not file) (cdr file)))))
+
 (define (valid-stack? stack)
   (if (null? stack)
       #t
@@ -23,5 +30,7 @@
        (assq 'procedure-name (car stack))
        (assq 'arguments (car stack))
        (assq 'environment (car stack))
-       (assq 'source (car stack))
+       (let ((source (assq 'source (car stack))))
+         (or (not source)
+             (valid-source? (cdr source))))
        (valid-stack? (cdr stack)))))